+++ NEWS.md
@@ -9,6 +9,7 @@
 The latest interface to optimizers in tensorflow are now used, these changes are described.
 
 * `gradient_descent` gains `momentum` and `nesterov` arguments, as described here: https://www.tensorflow.org/api_docs/python/tf/keras/optimizers/SGD
+* `adagrad` gains `epsilon` argument: https://www.tensorflow.org/api_docs/python/tf/keras/optimizers/Adagrad
 * 
 
 ## Minor